Metadata history Post Added

+++ TASC/webgui/views.py
@@ -670,37 +670,77 @@ def reverseimage(request):
             return render(request, 'apps/reverseimage.html',{"Error":"Do Select the File"})
 
 def metadata(request):
+  username = request.user.username
+  user = User.objects.filter(username=username).first()
+  
+  try:
+      with open("media/json/history_{}.json".format(username),"r") as file:
+        history=json.loads(file.read())
+        file.close()
+  except:
+    with open("media/json/history_{}.json".format(username),"w") as file:
+      history=json.loads(open("templates/json/history.json").read())
+      file.write(json.dumps(history, indent = 4))
+      file.close()
+  
   if request.method=="GET":
   
-    return render(request, 'apps/metadata.html',{"GET":'get'})
+    return render(request, 'apps/metadata.html',{"GET":'get',"history":history["app_type"]["metadata"][:10]})
 
   elif request.method=="POST":
+      
+    if 'metaimage' in request.FILES.keys() or "filename" in request.POST.keys():
+        if request.POST["filename"]!="" or request.FILES['metaimage'] != '' :
+            try:
+              filename=str(request.FILES['metaimage']).split('.',1)
+            except:
+              filename=request.POST["filename"].split('.',1)
 
-    username = request.user.username
-    user = User.objects.filter(username=username).first()
-        
-    if 'metaimage' in request.FILES.keys():
-        if request.FILES['metaimage'] != '':
-            filename=str(request.FILES['metaimage']).split('.',1)
             if len(filename)==2:
                 if filename[-1] in ['jpg','png','gif','tif','jpeg']:
+                  try:
+                    with open("media/metadata/{}.json".format(username),"r") as file:
+                      metafiles = json.loads(file.read())
+                  except:
+                    with open("media/metadata/{}.json".format(username),"w") as file:
+                      metafiles={}
+                      file.write(json.dumps({}))
+                  filename = ".".join(filename)
+                  
+                  if 'metaimage' in request.FILES.keys() and str(request.FILES['metaimage']) in metafiles.keys():
+                    metadata=metafiles[str(request.FILES['metaimage'])]
+                  elif "filename" in request.POST.keys() and filename in metafiles.keys():
+                    metadata=metafiles[str(filename)]
+                  else:
                     user.profile.metaimage = request.FILES['metaimage']
                     user.profile.save()
+                    history["app_type"]["metadata"].insert(0,{"filename":str(request.FILES['metaimage']),"time":datetime.now().astimezone(tz.gettz('ITC')).strftime('%H:%M %d %b')})
+                    with open("media/json/history_{}.json".format(username),"w") as file:
+                      file.write(json.dumps(history, indent = 4))
                     metaimage = user.profile.metaimage.url
                     googlemapapikey = user.profile.googlemapapikey
                     metadata = get_exif(metaimage)
-                    print(metadata)
+                    for i in metadata.keys():
+                      try:
+                        metadata[i]=str(str(metadata[i],"latin-1").replace("<"," ").replace(">"," "))
+                      except:
+                        pass
                     os.remove(BASE_DIR + user.profile.metaimage.url)
-                    if 'Error' in metadata.keys():
-                        return render(request, 'apps/metadata.html',{'metadata':metadata, "POST":"post"})
-                    elif 'Latitude' in metadata.keys():
-                          
-                        lats = metadata['Latitude']
-                        lons = metadata['Longitude']
-                        gmap3=heat_map([lats],[lons], googlemapapikey)
-                        return render(request, 'apps/metadata.html',{'metadata':metadata, 'gmap3':gmap3, "POST":"post"})
-                    else:
-                        return render(request, 'apps/metadata.html',{'metadata':metadata, "POST":"post"})
+                    metafiles[str(request.FILES['metaimage'])]=metadata
+
+                    with open("media/metadata/{}.json".format(username),"w") as file:
+                      file.write(json.dumps(metafiles))
+                    
+                  if 'Error' in metadata.keys():
+                      return render(request, 'apps/metadata.html',{'metadata':metadata, "POST":"post","history":history["app_type"]["metadata"][:10]})
+                  elif 'Latitude' in metadata.keys():
+                        
+                      lats = metadata['Latitude']
+                      lons = metadata['Longitude']
+                      gmap3=heat_map([lats],[lons], googlemapapikey)
+                      return render(request, 'apps/metadata.html',{'metadata':metadata, 'gmap3':gmap3, "POST":"post","history":history["app_type"]["metadata"][:10]})
+                  else:
+                      return render(request, 'apps/metadata.html',{'metadata':metadata, "POST":"post","history":history["app_type"]["metadata"][:10]})
                 elif filename[-1] == 'pdf':
                     user.profile.metaimage = request.FILES['metaimage']
                     user.profile.save()
@@ -708,7 +748,7 @@ def metadata(request):
                     os.remove(BASE_DIR + user.profile.metaimage.url)
                     metadata=pdf.get_metadata()
                     metadata['references_dict'] = pdf.get_references_as_dict()
-                    return render(request, 'apps/metadata.html',{'metadata':metadata, "POST":"post"})
+                    return render(request, 'apps/metadata.html',{'metadata':metadata, "POST":"post","history":history["app_type"]["metadata"][:10]})
                 else:
                     return render(request, 'apps/metadata.html',{"Error":"Upload a filename with Valid Extension"})
             else:
